Crate astros_nix

astros-nix is the Astros kernel, a safe, efficient unix-like operating system kernel built on top of KSTD and KSDK.

Modules

arch

context

The context that can be accessed from the current task, thread or process.

cpu

device

driver

error

events

fs

ipc

kcmdline

The module to parse kernel command-line arguments.

net

prelude

syscall

Read the Cpu ctx content then dispatch syscall to corresponding handler The each sub module contains functions that handle real syscall logic.

thread

Posix thread implementation

time

vm

Virtual memory (VM).

Macros

current

return current process

current_thread

Returns the current thread.

current_userspace

Gets the CurrentUserSpace from the current task.

impl_raw_sock_option_get_only

Impl RawSocketOption for a struct which is for only getsockopt and implements SocketOption.

impl_raw_socket_option

Impl RawSocketOption for a struct which implements SocketOption.

impl_socket_options

log_syscall_entry

match_sock_option_mut

match_sock_option_ref

return_errno

return_errno_with_message

Functions

init

main