astros_nix::context

Trait ReadCString

pub trait ReadCString {
    // Required method
    fn read_cstring(&mut self) -> Result<CString, Error>;
}

A trait providing the ability to read a C string from the user space.

The user space should be of the current process. The implemented method should read the bytes iteratively in the reader (VmReader) until encountering the end of the reader or reading a \0 (which is also included in the final C String).

Required Methods

fn read_cstring(&mut self) -> Result<CString, Error>