astros_nix

Module device

Modules

tty

Structs

PtyMaster

Pseudo terminal master. Internally, it has two buffers. One is inside ldisc, which is written by master and read by slave, the other is a ring buffer, which is written by slave and read by master.

PtySlave

Random

Urandom

Functions

get_device

init

Init the device node in fs, must be called after mounting rootfs.

new_pty_pair