astros_nix::device

Module tty

Modules

driver

line_discipline

termio

Structs

Tty

TtyDevice

Corresponds to /dev/tty in the file system. This device represents the controlling terminal of the session of current process.

Functions

get_n_tty

new_job_control_and_ldisc

open_ntty_as_controlling_terminal

Open N_TTY as the controlling terminal for the process. This method should only be called when creating the init process.