astros_nix::events

Trait EventsFilter

pub trait EventsFilter<E: Events>:
    Send
    + Sync
    + 'static {
    // Required method
    fn filter(&self, event: &E) -> bool;
}

A trait to filter events.

The no-op event filter

The unit type () can serve as a no-op event filter. It implements EventsFilter<E> for any events type E, with a filter method that always returns true. If the F type of Subject<E, F> is not specified explicitly, then the unit type () is chosen as the event filter.

Per-object event filter

Any Option<F: EventsFilter> is also an event filter thanks to the blanket implementations the EventsFilter trait. By using Option<F: EventsFilter>, we can decide, on a per-observer basis, if an observer needs an event filter.

Required Methods

fn filter(&self, event: &E) -> bool