astros_nix::fs

Module path

Form file paths within and across FSes with dentries and mount points.

Structs

Dentry

A Dentry is used to represent a location in the mount tree.

DentryKey

DentryKey is the unique identifier for the corresponding Dentry_.

MountNode

The MountNode is used to form a mount tree to maintain the mount information.

Functions

is_dot

Checks if the file name is “.”, indicating it’s the current directory.

is_dot_or_dotdot

Checks if the file name is “.” or “..”, indicating it’s the current or parent directory.

is_dotdot

Checks if the file name is “..”, indicating it’s the parent directory.