astros_nix::fs

Module utils

VFS components

Structs

Channel

A unidirectional communication channel, intended to implement IPC, e.g., pipe, unix domain sockets, etc.

Consumer

CreationFlags

Extension

An extension is a set of objects that is attached to an inode.

FileCreationMask

A mask for the file mode of a newly-created file or directory.

FileRange

A range in a file.

FixedCStr

An owned C-compatible string with a fixed capacity of N.

FixedStr

An owned string with a fixed capacity of N.

FlockItem

Represents a Flock item that can be held in a list of file locks. Each FlockItem contains a lock and a wait queue for threads that are blocked by the lock.

FlockList

Represents a list of non-POSIX file advisory locks (FLOCK). The list is used to manage file locks and resolve conflicts between them.

FsFlags

InodeMode

Metadata

PageCache

Permission

Producer

RangeLockItem

Represents a POSIX advisory file range lock in the kernel. Contains metadata about the lock and the processes waiting for it. The lock is associated with a specific range of the file.

RangeLockItemBuilder

Builder for RangeLockItem.

RangeLockList

List of File POSIX advisory range locks.

StatusFlags

SuperBlock

XattrName

Represents the name of an xattr. It includes both a valid namespace and a full name string slice, which contains the namespace prefix.

XattrSetFlags

Flags for setting an xattr value.

Enums

AccessMode

FallocMode

Represents the various operation modes for fallocate.

FlockType

Represents the type of a Flock - either shared or exclusive.

InodeType

IoctlCmd

MknodType

RangeLockType

Type of file range lock, aligned with Linux kernel. F_RDLCK = 0, F_WRLCK = 1, F_UNLCK = 2,

SeekFrom

XattrNamespace

Represents different namespaces with different capabilities for extended attributes (xattrs).

Constants

NAME_MAX

Maximum bytes in a file name

OFFSET_MAX

The maximum offset in a file.

PATH_MAX

Maximum bytes in a path

SYMLINKS_MAX

The upper limit for resolving symbolic links

XATTR_LIST_MAX_LEN

XATTR_NAME_MAX_LEN

XATTR_VALUE_MAX_LEN

Traits

DirEntryVecExt

DirentVisitor

A visitor for dir entries.

FileSystem

Inode

PageCacheBackend

This trait represents the backend for the page cache.

Functions

generate_random_operation

new_fs_in_memory

Type Aliases

CStr256

CachePage

A page in the page cache.

Str16

Str64