astros_nix::fs::utils

Trait DirEntryVecExt

pub trait DirEntryVecExt {
    // Required methods
    fn put_entry_if_not_found(
        &mut self,
        name: &str,
        f: impl Fn() -> Arc<dyn Inode>,
    );
    fn remove_entry_by_name(
        &mut self,
        name: &str,
    ) -> Option<(String, Arc<dyn Inode>)>;
}

Required Methods

fn put_entry_if_not_found(&mut self, name: &str, f: impl Fn() -> Arc<dyn Inode>)

If the entry is not found by name, use f to get the inode, then put the entry into vec.

fn remove_entry_by_name( &mut self, name: &str, ) -> Option<(String, Arc<dyn Inode>)>

Remove and returns the entry by name. Returns None if the entry has been removed.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.