astros_nix::fs::utils

Trait DirentVisitor

pub trait DirentVisitor {
    // Required method
    fn visit(
        &mut self,
        name: &str,
        ino: u64,
        type_: InodeType,
        offset: usize,
    ) -> Result<(), Error>;
}

A visitor for dir entries.

Required Methods

fn visit( &mut self, name: &str, ino: u64, type_: InodeType, offset: usize, ) -> Result<(), Error>

Visit a dir entry.

If the visitor succeeds in visiting the given inode, an Ok(()) is returned; Otherwise, an error is returned. Different implementations for DirentVisitor may choose to report errors for different reasons. Regardless of the exact errors and reasons, readdir-family methods shall stop feeding the visitor with the next inode as long as an error is returned by the visitor.

Example

Vec<String> is implemented as DirentVisitor so that the file names under a dir can be easily collected, which is convenient for testing purposes.

let mut all_dirents = Vec::new();
let dir_inode = todo!("create an inode");
dir_inode.readdir_at(0, &mut all_dirents).unwrap();