astros_nix::fs::utils

Trait Inode

pub trait Inode:
    Any
    + Sync
    + Send {

    // Required methods
    fn size(&self) -> usize;
    fn resize(&self, new_size: usize) -> Result<(), Error>;
    fn metadata(&self) -> Metadata;
    fn ino(&self) -> u64;
    fn type_(&self) -> InodeType;
    fn mode(&self) -> Result<InodeMode, Error>;
    fn set_mode(&self, mode: InodeMode) -> Result<(), Error>;
    fn owner(&self) -> Result<Uid, Error>;
    fn set_owner(&self, uid: Uid) -> Result<(), Error>;
    fn group(&self) -> Result<Gid, Error>;
    fn set_group(&self, gid: Gid) -> Result<(), Error>;
    fn atime(&self) -> Duration;
    fn set_atime(&self, time: Duration);
    fn mtime(&self) -> Duration;
    fn set_mtime(&self, time: Duration);
    fn ctime(&self) -> Duration;
    fn set_ctime(&self, time: Duration);
    fn fs(&self) -> Arc<dyn FileSystem>;

    // Provided methods
    fn page_cache(&self) -> Option<Vmo<Full>> { ... }
    fn read_at(
        &self,
        offset: usize,
        writer: &mut VmWriter<'_>,
    ) -> Result<usize, Error> { ... }
    fn read_direct_at(
        &self,
        offset: usize,
        writer: &mut VmWriter<'_>,
    ) -> Result<usize, Error> { ... }
    fn write_at(
        &self,
        offset: usize,
        reader: &mut VmReader<'_>,
    ) -> Result<usize, Error> { ... }
    fn write_direct_at(
        &self,
        offset: usize,
        reader: &mut VmReader<'_>,
    ) -> Result<usize, Error> { ... }
    fn create(
        &self,
        name: &str,
        type_: InodeType,
        mode: InodeMode,
    ) -> Result<Arc<dyn Inode>, Error> { ... }
    fn mknod(
        &self,
        name: &str,
        mode: InodeMode,
        type_: MknodType,
    ) -> Result<Arc<dyn Inode>, Error> { ... }
    fn as_device(&self) -> Option<Arc<dyn Device>> { ... }
    fn readdir_at(
        &self,
        offset: usize,
        visitor: &mut dyn DirentVisitor,
    ) -> Result<usize, Error> { ... }
    fn link(&self, old: &Arc<dyn Inode>, name: &str) -> Result<(), Error> { ... }
    fn unlink(&self, name: &str) -> Result<(), Error> { ... }
    fn rmdir(&self, name: &str) -> Result<(), Error> { ... }
    fn lookup(&self, name: &str) -> Result<Arc<dyn Inode>, Error> { ... }
    fn rename(
        &self,
        old_name: &str,
        target: &Arc<dyn Inode>,
        new_name: &str,
    ) -> Result<(), Error> { ... }
    fn read_link(&self) -> Result<String, Error> { ... }
    fn write_link(&self, target: &str) -> Result<(), Error> { ... }
    fn ioctl(&self, cmd: IoctlCmd, arg: usize) -> Result<i32, Error> { ... }
    fn sync_all(&self) -> Result<(), Error> { ... }
    fn sync_data(&self) -> Result<(), Error> { ... }
    fn fallocate(
        &self,
        mode: FallocMode,
        offset: usize,
        len: usize,
    ) -> Result<(), Error> { ... }
    fn poll(&self, mask: IoEvents, _poller: Option<&mut PollHandle>) -> IoEvents { ... }
    fn is_dentry_cacheable(&self) -> bool { ... }
    fn is_seekable(&self) -> bool { ... }
    fn extension(&self) -> Option<&Extension> { ... }
    fn set_xattr(
        &self,
        name: XattrName<'_>,
        value_reader: &mut VmReader<'_>,
        flags: XattrSetFlags,
    ) -> Result<(), Error> { ... }
    fn get_xattr(
        &self,
        name: XattrName<'_>,
        value_writer: &mut VmWriter<'_>,
    ) -> Result<usize, Error> { ... }
    fn list_xattr(
        &self,
        namespace: XattrNamespace,
        list_writer: &mut VmWriter<'_>,
    ) -> Result<usize, Error> { ... }
    fn remove_xattr(&self, name: XattrName<'_>) -> Result<(), Error> { ... }
    fn check_permission(&self, perm: Permission) -> Result<(), Error> { ... }

}

Required Methods

fn size(&self) -> usize

fn resize(&self, new_size: usize) -> Result<(), Error>

fn metadata(&self) -> Metadata

fn ino(&self) -> u64

fn type_(&self) -> InodeType

fn mode(&self) -> Result<InodeMode, Error>

fn set_mode(&self, mode: InodeMode) -> Result<(), Error>

fn owner(&self) -> Result<Uid, Error>

fn set_owner(&self, uid: Uid) -> Result<(), Error>

fn group(&self) -> Result<Gid, Error>

fn set_group(&self, gid: Gid) -> Result<(), Error>

fn atime(&self) -> Duration

fn set_atime(&self, time: Duration)

fn mtime(&self) -> Duration

fn set_mtime(&self, time: Duration)

fn ctime(&self) -> Duration

fn set_ctime(&self, time: Duration)

fn fs(&self) -> Arc<dyn FileSystem>

Provided Methods

fn page_cache(&self) -> Option<Vmo<Full>>

fn read_at( &self, offset: usize, writer: &mut VmWriter<'_>, ) -> Result<usize, Error>

fn read_direct_at( &self, offset: usize, writer: &mut VmWriter<'_>, ) -> Result<usize, Error>

fn write_at( &self, offset: usize, reader: &mut VmReader<'_>, ) -> Result<usize, Error>

fn write_direct_at( &self, offset: usize, reader: &mut VmReader<'_>, ) -> Result<usize, Error>

fn create( &self, name: &str, type_: InodeType, mode: InodeMode, ) -> Result<Arc<dyn Inode>, Error>

fn mknod( &self, name: &str, mode: InodeMode, type_: MknodType, ) -> Result<Arc<dyn Inode>, Error>

fn as_device(&self) -> Option<Arc<dyn Device>>

fn readdir_at( &self, offset: usize, visitor: &mut dyn DirentVisitor, ) -> Result<usize, Error>

fn link(&self, old: &Arc<dyn Inode>, name: &str) -> Result<(), Error>

fn unlink(&self, name: &str) -> Result<(), Error>

fn rmdir(&self, name: &str) -> Result<(), Error>

fn lookup(&self, name: &str) -> Result<Arc<dyn Inode>, Error>

fn rename( &self, old_name: &str, target: &Arc<dyn Inode>, new_name: &str, ) -> Result<(), Error>

fn read_link(&self) -> Result<String, Error>

fn write_link(&self, target: &str) -> Result<(), Error>

fn ioctl(&self, cmd: IoctlCmd, arg: usize) -> Result<i32, Error>

fn sync_all(&self) -> Result<(), Error>

fn sync_data(&self) -> Result<(), Error>

fn fallocate( &self, mode: FallocMode, offset: usize, len: usize, ) -> Result<(), Error>

Manipulates a range of space of the file according to the specified allocate mode, the manipulated range starts at offset and continues for len bytes.

fn poll(&self, mask: IoEvents, _poller: Option<&mut PollHandle>) -> IoEvents

fn is_dentry_cacheable(&self) -> bool

Returns whether a VFS dentry for this inode should be put into the dentry cache.

The dentry cache in the VFS layer can accelerate the lookup of inodes. So usually, it is preferable to use the dentry cache. And thus, the default return value of this method is true.

But this caching can raise consistency issues in certain use cases. Specifically, the dentry cache works on the assumption that all FS operations go through the dentry layer first. This is why the dentry cache can reflect the up-to-date FS state. Yet, this assumption may be broken. If the inodes of a file system may โdisappearโ without unlinking through the VFS layer, then their dentries should not be cached. For example, an inode in procfs (say, /proc/1/fd/2) can โdisappearโ without notice from the perspective of the dentry cache. So for such inodes, they are incompatible with the dentry cache. And this method returns false.

Note that if any ancestor directory of an inode has this method returns false, then this inode would not be cached by the dentry cache, even when the method of this inode returns true.

fn is_seekable(&self) -> bool

fn extension(&self) -> Option<&Extension>

Get the extension of this inode

fn set_xattr( &self, name: XattrName<'_>, value_reader: &mut VmReader<'_>, flags: XattrSetFlags, ) -> Result<(), Error>

fn get_xattr( &self, name: XattrName<'_>, value_writer: &mut VmWriter<'_>, ) -> Result<usize, Error>

fn list_xattr( &self, namespace: XattrNamespace, list_writer: &mut VmWriter<'_>, ) -> Result<usize, Error>

fn remove_xattr(&self, name: XattrName<'_>) -> Result<(), Error>

fn check_permission(&self, perm: Permission) -> Result<(), Error>

Used to check for read/write/execute permissions on a file.

Similar to Linux, using โfsuidโ here allows setting filesystem permissions without changing the โnormalโ uids for other tasks.

Implementations

impl dyn Inode

pub fn downcast_ref<T: Inode>(&self) -> Option<&T>

pub fn read_to_end(&self, buf: &mut Vec<u8>) -> Result<usize, Error>

pub fn read_direct_to_end(&self, buf: &mut Vec<u8>) -> Result<usize, Error>

pub fn writer(&self, from_offset: usize) -> InodeWriter<'_>

pub fn read_bytes_at( &self, offset: usize, buf: &mut [u8], ) -> Result<usize, Error>

pub fn write_bytes_at(&self, offset: usize, buf: &[u8]) -> Result<usize, Error>

pub fn read_bytes_direct_at( &self, offset: usize, buf: &mut [u8], ) -> Result<usize, Error>

pub fn write_bytes_direct_at( &self, offset: usize, buf: &[u8], ) -> Result<usize, Error>