astros_nix::ipc::semaphore::system_v

Module sem_set

Structs

SemaphoreSet

Constants

SEMAEM

Maximum value that can be recorded for semaphore adjustment (SEM_UNDO).

SEMMNI

Maximum number of semaphore sets.

SEMMNS

Maximum number of seaphores in all semaphore sets.

SEMMSL

Maximum number of semaphores per semaphore ID.

SEMOPM

Maximum number of operations for semop.

SEMVMX

MAximum semaphore value.

Functions

check_sem

Checks the semaphore. Return Ok if the semaphore exists and pass the check.

create_sem_set

create_sem_set_with_id

sem_sets

sem_sets_mut