astros_nix::thread

Module oops

Kernel “oops” handling.

In Astros, a Rust panic leads to a kernel “oops”. A kernel oops behaves as an exceptional control flow event. If kernel oopses happened too many times, the kernel panics and the system gets halted. Kernel oops are per- thread, so one thread’s oops does not affect other threads.

Though we can recover from the Rust panics. It is generally not recommended to make Rust panics as a general exception handling mechanism. Handling exceptions with Result is more idiomatic.

Structs

OopsInfo

The kernel “oops” information.

Functions

catch_panics_as_oops

Executes the given function and catches any panics that occur.