astros_nix::thread::work_queue::worker_pool

Trait WorkerScheduler

pub trait WorkerScheduler: Sync + Send {
    // Required method
    fn schedule(&self);
}

Schedule Workers for a WorkerPool.

Having an excessive number of Workers in WorkerPool may result in wastage of system resources, while a shortage of workers may lead to longer response time for workitems. A well-designed WorkerScheduler must strike a balance between resource utilization and response time.

Required Methods

fn schedule(&self)

Schedule workers in a worker pool. This needs to solve two problems: when to increase or decrease workers, and how to add or remove workers to keep the number of workers in a reasonable range.