astros_nix

Module time

Modules

clocks

timer

wait

Structs

SystemTime

This struct corresponds to SystemTime in Rust std.

Timer

A timer with periodic functionality.

TimerManager

TimerManager is used to create timers and manage their expiries. It holds a clock and can create Timers based on this clock.

UnixTime

Unix time measures time by the number of seconds that have elapsed since the Unix epoch, without adjustments made due to leap seconds.

itimerspec_t

This struct is corresponding to the itimerspec struct in Linux.

itimerval_t

This struct is corresponding to the itimerval struct in Linux.

timespec_t

timeval_t

Constants

TIMER_ABSTIME

The various flags for setting POSIX.1b interval timers:

Statics

START_TIME

Traits

Clock

A trait that can abstract clocks which have the ability to read time, and has a fixed resolution.

Type Aliases

clock_t

clockid_t

suseconds_t

time_t