astros_nix::time

Module clocks

Structs

BootTimeClock

BootTimeClock measures the time elapsed since the system was booted, including time when the system was suspended.

CpuClock

A clock used to record the CPU time for processes and threads.

JiffiesClock

The Clock that reads the jiffies, and turn the counter into Duration.

MonotonicClock

MonotonicClock represents a clock that measures time in a way that is monotonically increasing since the system was booted.

MonotonicCoarseClock

MonotonicCoarseClock is a coarse-grained version of the monotonic clock.

MonotonicRawClock

MonotonicRawClock provides raw monotonic time that is not influenced by NTP corrections.

ProfClock

A profiling clock that contains a user CPU clock and a kernel CPU clock.

RealTimeClock

RealTimeClock represents a clock that provides the current real time.

RealTimeCoarseClock

RealTimeCoarseClock is a coarse-grained version of a real-time clock.

Statics

CLOCK_BOOTTIME_INSTANCE

CLOCK_BOOTTIME_MANAGER

CLOCK_MONOTONIC_COARSE_INSTANCE

CLOCK_MONOTONIC_INSTANCE

CLOCK_MONOTONIC_MANAGER

CLOCK_MONOTONIC_RAW_INSTANCE

CLOCK_REALTIME_COARSE_INSTANCE

CLOCK_REALTIME_INSTANCE

CLOCK_REALTIME_MANAGER

JIFFIES_TIMER_MANAGER

The system-wide TimerManager for the JiffiesClock.