astros_nix::time

Trait Clock

pub trait Clock: Send + Sync {
    // Required method
    fn read_time(&self) -> Duration;

    // Provided method
    fn resolution(&self) -> u64
       where Self: Sized { ... }
}

A trait that can abstract clocks which have the ability to read time, and has a fixed resolution.

Required Methods

fn read_time(&self) -> Duration

Read the current time of this clock.

Provided Methods

fn resolution(&self) -> u64
where Self: Sized,

Return the resolution of this clock. Set to the resolution of system time interrupt by default.