astros_nix

Module vm

Virtual memory (VM).

There are two primary VM abstractions:

Virtual Memory Address Regions (VMARs) a type of capability that manages user address spaces.

Virtual Memory Objects (VMOs) are are a type of capability that represents a set of memory pages.

The concepts of VMARs and VMOs are originally introduced by Zircon. As capabilities, the two abstractions are aligned with our goal of everything-is-a-capability, although their specifications and implementations in C/C++ cannot apply directly to Astros. In Astros, VMARs and VMOs, as well as other capabilities, are implemented as zero-cost capabilities.

Modules

page_fault_handler

perms

util

vmar

Virtual Memory Address Regions (VMARs).

vmo

Virtual Memory Objects (VMOs).

Functions

mem_total

Total physical memory in the entire system in bytes.