astros_nix::vm::page_fault_handler

Trait PageFaultHandler

pub trait PageFaultHandler {
    // Required method
    fn handle_page_fault(
        &self,
        page_fault_info: &PageFaultInfo,
    ) -> Result<(), Error>;
}

This trait is implemented by structs which can handle a user space page fault.

Required Methods

fn handle_page_fault( &self, page_fault_info: &PageFaultInfo, ) -> Result<(), Error>

Handle a page fault, whose information is provided in page_fault_info.

Returns Ok if the page fault is handled successfully, Err otherwise.