astros_nix::vm

Module vmar

Virtual Memory Address Regions (VMARs).

Modules

vm_mapping

Structs

Vmar

Virtual Memory Address Regions (VMARs) are a type of capability that manages user address spaces.

VmarMapOptions

Options for creating a new mapping. The mapping is not allowed to overlap with any child VMARs. And unless specified otherwise, it is not allowed to overlap with any existing mapping, either.

Constants

ROOT_VMAR_LOWEST_ADDR

Traits

VmarRightsOp

Functions

get_intersected_range

Gets the intersection range of two ranges. The two ranges should be ensured to be intersected.

is_intersected

Determines whether two ranges are intersected. returns false if one of the ranges has a length of 0

is_userspace_vaddr

Returns whether the input vaddr is a legal user space virtual address.