astros_nix::vm

Module vmo

Virtual Memory Objects (VMOs).

Structs

CommitFlags

Commit Flags.

Vmo

Virtual Memory Objects (VMOs) are a type of capability that represents a range of memory pages.

VmoFlags

VMO flags.

VmoOptions

Options for allocating a root VMO.

Traits

Pager

Pagers provide frame to a VMO.

VmoRightsOp

Functions exist both for static capbility and dynamic capability

Functions

get_page_idx_range

Gets the page index range that contains the offset range of VMO.